feat : 어드민 리뷰 승인 기능 구현 (#62)
* feat : default 프로필 src 변경

* feat : getAuthority 함수 수정

* feat : admin API Role_ADMIN만 접근 가능하도록 변경

* feat : update Status, RejectReason 함수 추가

* feat : 리뷰 승인 요청 Dto 추가

* feat : 어드민 리뷰 승인 API 추가

* feat : 어드민 리뷰 승인 Service 함수 구현

* feat : 어드민 리뷰 승인 log 출력을 위한 Aop추가

* chore : 함수 오타 수정

* chore : 함수 오타 수정

* fix : reviewCount Integer 형으로 변경

+++ concertseat/src/main/java/site/concertseat/domain/review/repository/ReviewRepository.java
@@ -44,7 +44,7 @@ public interface ReviewRepository extends JpaRepository<Review, Long>, CustomRev
     @Query("select count(r) " +
             "from Review r " +
             "where r.status = :status")
-    Long countApprovedReviews(@Param("status") ReviewStatus status);
+    Integer countApprovedReviews(@Param("status") ReviewStatus status);
 
     @Query("select count(r) " +
             "from Review r " +
